Add link to specific yq package repository
There are two widely-used versions of `yq`, one by [mikefarah](https://github.com/mikefarah/yq) and another by [kislyuk](https://github.com/kislyuk/yq). This adds a link to the repo of the version of `yq` that is being used in this image to make it clear which version is being used.